+++ packages/config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @m5kdev/config
 
+## 0.8.2
+
+### Patch Changes
+
+- types
+
+## 0.8.1
+
+### Patch Changes
+
+- exports
+
+## 0.8.0
+
+### Minor Changes
+
+- tsdown
+
 ## 0.7.0
 
 ### Minor Changes