+++ packages/web-ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @m5kdev/web-ui
 
+## 0.8.2
+
+### Patch Changes
+
+- types
+- Updated dependencies
+  - @m5kdev/frontend@0.8.2
+  - @m5kdev/commons@0.8.2
+
+## 0.8.1
+
+### Patch Changes
+
+- exports
+- Updated dependencies
+  - @m5kdev/frontend@0.8.1
+  - @m5kdev/commons@0.8.1
+
+## 0.8.0
+
+### Minor Changes
+
+- tsdown
+
+### Patch Changes
+
+- Updated dependencies
+  - @m5kdev/frontend@0.8.0
+  - @m5kdev/commons@0.8.0
+
 ## 0.7.0
 
 ### Minor Changes